Change health from Integers to Float values

+++ ProjectSnow/Assets/_Scripts/Damage System/Attack/Attack.cs
@@ -29,7 +29,7 @@ namespace Game.DamageSystem.Attacks
         
         #region Damage Settings
         [FoldoutGroup("Damage Settings")]
-        [SerializeField] protected int DamageAmount;
+        [SerializeField] protected float DamageAmount;
         
         [FoldoutGroup("Damage Settings")]
         [SerializeField] protected LayerMask Layers;
